requirements.txt
# =============================================================================
# Instant Policy - Requirements
# Paper: "Instant Policy: In-Context Imitation Learning via Graph Diffusion"
# =============================================================================
#
# Installation Guide:
# -------------------
# 1. First install PyTorch with CUDA (choose one):
#
#    For CUDA 11.8:
#    pip install torch==2.2.0 torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
#
#    For CUDA 12.1:
#    pip install torch==2.2.0 torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
#
# 2. Then install PyTorch Geometric (match your torch+cuda version):
#
#    For torch 2.2.0 + CUDA 11.8:
#    pip install torch-scatter torch-cluster -f https://data.pyg.org/whl/torch-2.2.0+cu118.html
#    pip install torch-geometric==2.5.0
#
#    For torch 2.2.0 + CUDA 12.1:
#    pip install torch-scatter torch-cluster -f https://data.pyg.org/whl/torch-2.2.0+cu121.html
#    pip install torch-geometric==2.5.0
#
# 3. Finally install the rest:
#    pip install -r requirements.txt
#
# =============================================================================

# --- Core Deep Learning (install separately, see above) ---
# torch>=2.2.0
# torch-geometric>=2.5.0
# torch-scatter
# torch-cluster

# --- Training Framework ---
pytorch-lightning>=2.4.0
torchmetrics>=1.0.0

# --- Point Cloud Processing ---
open3d>=0.18.0

# --- Scientific Computing ---
numpy>=1.26.0
scipy>=1.14.0

# --- Configuration & Utilities ---
pyyaml>=6.0
tqdm>=4.66.0

# --- Optional: Visualization & Logging ---
matplotlib>=3.9.0
wandb>=0.18.0

# --- Optional: Development ---
# pytest
# black
# flake8
